# swagger.yaml
openapi: 3.0.0
info:
    title: 'API Documentation'
    version: 1.0.0
    description: 'API documentation for plans, workouts, and exercises.'
servers:
    -
        url: 'http://localhost:3010'
paths:
...
    /plans:
        post:
            summary: 'Create a new training plan'
            tags: [Plans]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {title: {type: string, example: 'Beginner Workout Plan'}, length: {type: integer, example: 30}, coach: {type: string, example: 'John Doe'}, description: {type: string, example: 'A comprehensive plan for beginners.'}}}}}}
            responses: {'201': {description: 'Plan created successfully', content: {application/json: {schema: {type: object, properties: {id: {type: integer, example: 1}, title: {type: string, example: 'Beginner Workout Plan'}, length: {type: integer, example: 30}, coach: {type: string, example: 'John Doe'}, description: {type: string, example: 'A comprehensive plan for beginners.'}}}}}}, '400': {description: 'Missing or invalid fields'}, '403': {description: 'Forbidden, if the user does not have a valid role to create the plan'}, '500': {description: 'Server error'}}
        get:
            summary: 'Retrieve a list of training plans'
            tags: [Plans]
            responses: {'200': {description: 'A list of training plans', content: {application/json: {schema: {type: array, items: {type: object, properties: {id: {type: integer, example: 1}, title: {type: string, example: 'Beginner Workout Plan'}, length: {type: integer, example: 30}, coach: {type: string, example: 'John Doe'}, description: {type: string, example: 'A comprehensive plan for beginners.'}}}}}}}, '403': {description: 'Forbidden. You do not have the required permissions.'}}
    '/plans/{id}':
        get:
            summary: 'Retrieve a specific training plan by ID'
            tags: [Plans]
            parameters: [{name: id, in: path, required: true, description: 'ID of the training plan', schema: {type: integer}}]
            responses: {'200': {description: 'A training plan found', content: {application/json: {schema: {type: object, properties: {id: {type: integer, example: 1}, title: {type: string, example: 'Beginner Workout Plan'}, length: {type: integer, example: 30}, coach: {type: string, example: 'John Doe'}, description: {type: string, example: 'A comprehensive plan for beginners.'}}}}}}, '404': {description: 'Plan not found'}}
        put:
            summary: 'Update a specific training plan'
            tags: [Plans]
            parameters: [{name: id, in: path, required: true, description: 'ID of the training plan', schema: {type: integer}}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {title: {type: string, example: 'Beginner Workout Plan'}, length: {type: integer, example: 30}, coach: {type: string, example: 'John Doe'}, description: {type: string, example: 'A comprehensive plan for beginners.'}}}}}}
            responses: {'200': {description: 'Plan updated successfully'}, '400': {description: 'All fields are required'}, '403': {description: 'Forbidden. You do not have the required permissions to update this plan.'}, '404': {description: 'Plan not found'}, '422': {description: 'Invalid input values'}}
        delete:
            summary: 'Delete a specific training plan'
            tags: [Plans]
            parameters: [{name: id, in: path, required: true, description: 'ID of the training plan', schema: {type: integer}}]
            responses: {'200': {description: 'Plan deleted successfully'}, '403': {description: 'Access denied. You do not have permission to delete this plan.'}, '404': {description: 'Plan not found'}}
    '/plans/{plan_id}/workouts':
        get:
            summary: 'Retrieve all workouts for a specific plan'
            tags: [Plans]
            parameters: [{name: plan_id, in: path, required: true, description: 'ID of the training plan', schema: {type: integer}}]
            responses: {'200': {description: 'A list of workouts for the specified plan', content: {application/json: {schema: {type: array, items: {type: object, properties: {id: {type: integer, example: 1}, plan_id: {type: integer, example: 1}, name: {type: string, example: 'Morning Cardio'}, length: {type: integer, example: 30}, type: {type: string, example: Cardio}, frequency: {type: integer, example: 3}}}}}}}, '403': {description: 'Access denied. Only users and admins can access this endpoint.'}, '404': {description: 'Plan not found'}}
...